Use float16 on cuda device

+++ GraphRec-WWW19/Social_Aggregators.py
@@ -18,7 +18,7 @@ class Social_Aggregator(nn.Module):
         self.att = Attention(self.embed_dim)
 
     def forward(self, nodes, to_neighs):
-        embed_matrix = torch.empty(len(nodes), self.embed_dim, dtype=torch.float).to(self.device)
+        embed_matrix = torch.empty(len(nodes), self.embed_dim).to(self.device, dtype=torch.float16).float()
         for i in range(len(nodes)):
             tmp_adj = to_neighs[i]
             num_neighs = len(tmp_adj)
